JWWEB-737 Split component Added description for statuses
